+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+## [1.30.5](https://github.com/Ironsail-llc/genus-os/compare/v1.30.4...v1.30.5) (2026-08-19)
+
+### Bug Fixes
+
+* **dist:** installer resolves paths outside site-packages installs ([#248](https://github.com/Ironsail-llc/genus-os/issues/248)) ([7ae1933](https://github.com/Ironsail-llc/genus-os/commit/7ae19335548af3a2516cf9d907bc41535865e31e)), closes [#245](https://github.com/Ironsail-llc/genus-os/issues/245)
+
+## [1.30.4](https://github.com/Ironsail-llc/genus-os/compare/v1.30.3...v1.30.4) (2026-08-19)
+
+### Bug Fixes
+
+* **config:** load the workspace .env that init writes ([#247](https://github.com/Ironsail-llc/genus-os/issues/247)) ([40693c2](https://github.com/Ironsail-llc/genus-os/commit/40693c2f817652f634639d47289830f6a939f89c))
+
+## [1.30.3](https://github.com/Ironsail-llc/genus-os/compare/v1.30.2...v1.30.3) (2026-08-19)
+
+### Bug Fixes
+
+* **docker:** copy the scaffold templates into the python image build ([#254](https://github.com/Ironsail-llc/genus-os/issues/254)) ([376e2ad](https://github.com/Ironsail-llc/genus-os/commit/376e2adf870c7d193cc397da65874c97546ef684)), closes [#250](https://github.com/Ironsail-llc/genus-os/issues/250)
+
+## [1.30.2](https://github.com/Ironsail-llc/genus-os/compare/v1.30.1...v1.30.2) (2026-08-18)
+
+### Bug Fixes
+
+* **cli:** init next-steps match the install mode ([#246](https://github.com/Ironsail-llc/genus-os/issues/246)) ([0113234](https://github.com/Ironsail-llc/genus-os/commit/0113234ec3da709ce1a1471288b4d10df8794591))
+* **docker:** include the scaffold templates in the build context ([#250](https://github.com/Ironsail-llc/genus-os/issues/250)) ([0b47423](https://github.com/Ironsail-llc/genus-os/commit/0b47423340c2a44ee2420394479eaad481c22d30)), closes [#245](https://github.com/Ironsail-llc/genus-os/issues/245) [245/#246](https://github.com/245/genus-os/issues/246)
+
+## [1.30.1](https://github.com/Ironsail-llc/genus-os/compare/v1.30.0...v1.30.1) (2026-08-18)
+
+### Bug Fixes
+
+* **dist:** ship the init scaffold in the wheel via package data ([#245](https://github.com/Ironsail-llc/genus-os/issues/245)) ([0e32135](https://github.com/Ironsail-llc/genus-os/commit/0e321357b40430f701f9c238c73c7f487bd08378))
+
 ## [1.30.0](https://github.com/Ironsail-llc/genus-os/compare/v1.29.2...v1.30.0) (2026-08-18)
 
 ### Features